--- a/server/src/test/java/org/opensearch/index/SlowLogLevelTests.java
+++ b/server/src/test/java/org/opensearch/index/SlowLogLevelTests.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.index;
 
+import org.opensearch.common.logging.SlowLogLevel;
 import org.opensearch.test.OpenSearchTestCase;
 
 public class SlowLogLevelTests extends OpenSearchTestCase {