@msfroh 2023-11-06
Hey -- sorry for the late idea, but can we maybe implement this a completely different way: as a search pipeline response processor?

It occurs to me that things like this that run at the coordinator level and depend on the search request/response could use search pipelines, rather than needing to modify OpenSearch core at all.