@dzane17 2023-11-09
@msfroh Thanks for bringing up search response processors. I believe it can be done via search pipelines and I get your point about shifting logic into modules, but it does not exactly fit our use case. Coordinator slow logs are designed similar to existing shard slow logs so it does not make sense to implement then in different ways. I could see the argument if we wanted to move both into search pipelines (although don't think processors can be injected at the shard level currently).

Also, we want to keep the user experience consistent between request vs shard slow logs. For search pipelines, users would need to pass a parameter with every request to enable logging. This itself is not an issue, but seems unintuitive for request slow logs to require a param while shard slow logs uses index settings.
```
// To use a pipeline with a query, specify the pipeline name in the search_pipeline query parameter:
GET /my_index/_search?search_pipeline=my_pipeline
```