@msfroh 2023-11-09
> For search pipelines, users would need to pass a parameter with every request to enable logging.

That's not actually true. You can specify a default search pipeline for an index, such that the pipeline gets applied on every request (by default).

> Coordinator slow logs are designed similar to existing shard slow logs so it does not make sense to implement then in different ways.

That's kind of a circular argument: "We did it this way, so it doesn't make sense to do it another way."

I would argue that the existing shard-level slow logs were not a great choice in the first place. Why would I want shard-level logging for slow queries before having request-level logging for slow queries? 

In my opinion, we have an opportunity to build something that is more useful (i.e. request-level logging), easier to configure (i.e. a response processor), and that doesn't need to be permanently woven into core search APIs.