@msfroh 2023-11-10
So, I've been thinking about this some more, and I think I like both solutions:

1. coordinator slow logs that behave like shard-level slow logs as a short-term solution that addresses the (significant) problems with shard level slow logs (see e.g. https://github.com/opensearch-project/OpenSearch/issues/2054), and
2. a search pipeline response processor that can log queries, optionally based on conditions.

IMO, we should do both. 

The response processor query logging doesn't need to be limited to slow logging. (Maybe I only want to log every query containing a prefix query. Maybe I want to log different subsets of queries to different places.) In the long run, I think the response processor logging could replace coordinator-level slow logging, but that should be up to users to decide which they want to use.