@msfroh 2023-11-15
Looks like you have a merge conflict still. 

The `buildRewriteListener` method still uses `searchRequestOperationsListener`:

```
> Task :server:compileJava FAILED
/var/jenkins/workspace/gradle-check/search/server/src/main/java/org/opensearch/action/search/TransportSearchAction.java:525: error: cannot find symbol
                searchRequestOperationsListener
                ^
  symbol:   variable searchRequestOperationsListener
  location: class TransportSearchAction
/var/jenkins/workspace/gradle-check/search/server/src/main/java/org/opensearch/action/search/TransportSearchAction.java:579: error: cannot find symbol
                    searchRequestContext
                    ^
  symbol:   variable searchRequestContext
  location: class TransportSearchAction
/var/jenkins/workspace/gradle-check/search/server/src/main/java/org/opensearch/action/search/TransportSearchAction.java:601: error: cannot find symbol
                            searchRequestContext
                            ^
  symbol:   variable searchRequestContext
  location: class TransportSearchAction
/var/jenkins/workspace/gradle-check/search/server/src/main/java/org/opensearch/action/search/TransportSearchAction.java:652: error: cannot find symbol
                                searchRequestContext
```